Refactor input classes into their own files
