+++ src/electionguard_cli/cli_models/__init__.py
@@ -10,9 +10,9 @@ from electionguard_cli.cli_models.e2e_decrypt_results import (
     E2eDecryptResults,
 )
 from electionguard_cli.cli_models.e2e_inputs import (
+    CliElectionInputsBase,
     E2eInputs,
     ImportBallotInputs,
-    CliElectionInputsBase,
 )
 from electionguard_cli.cli_models.e2e_submit_results import (
     E2eSubmitResults,
@@ -20,10 +20,10 @@ from electionguard_cli.cli_models.e2e_submit_results import (
 
 __all__ = [
     "BuildElectionResults",
+    "CliElectionInputsBase",
     "E2eDecryptResults",
     "E2eInputs",
     "E2eSubmitResults",
-    "CliElectionInputsBase",
     "ImportBallotInputs",
     "e2e_build_election_results",
     "e2e_decrypt_results",
