+++ src/electionguard_cli/e2e_steps/__init__.py
@@ -2,7 +2,6 @@ from electionguard_cli.e2e_steps import decrypt_step
 from electionguard_cli.e2e_steps import e2e_step_base
 from electionguard_cli.e2e_steps import election_builder_step
 from electionguard_cli.e2e_steps import election_record_step
-from electionguard_cli.e2e_steps import input_retrieval_step
 from electionguard_cli.e2e_steps import key_ceremony_step
 from electionguard_cli.e2e_steps import print_results_step
 from electionguard_cli.e2e_steps import submit_votes_step
@@ -19,11 +18,6 @@ from electionguard_cli.e2e_steps.election_builder_step import (
 from electionguard_cli.e2e_steps.election_record_step import (
     ElectionRecordStep,
 )
-from electionguard_cli.e2e_steps.input_retrieval_step import (
-    E2eInputRetrievalStep,
-    ImportBallotsInputRetrievalStep,
-    InputRetrievalStepBase,
-)
 from electionguard_cli.e2e_steps.key_ceremony_step import (
     KeyCeremonyStep,
 )
@@ -36,12 +30,9 @@ from electionguard_cli.e2e_steps.submit_votes_step import (
 
 __all__ = [
     "DecryptStep",
-    "E2eInputRetrievalStep",
     "E2eStepBase",
     "ElectionBuilderStep",
     "ElectionRecordStep",
-    "ImportBallotsInputRetrievalStep",
-    "InputRetrievalStepBase",
     "KeyCeremonyStep",
     "PrintResultsStep",
     "SubmitVotesStep",
@@ -49,7 +40,6 @@ __all__ = [
     "e2e_step_base",
     "election_builder_step",
     "election_record_step",
-    "input_retrieval_step",
     "key_ceremony_step",
     "print_results_step",
     "submit_votes_step",
